move scrubadub.detectors.stanford_ner -> scrubadub.detectors.stanford
