+++ scrubadub/detectors/stanford.py
